+++ decidim-blogs/db/migrate/20200128094730_add_endorsements_counter_cache_to_blogs.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddEndorsementsCounterCacheToBlogs < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_blogs_posts, :endorsements_count, :integer, null: false, default: 0
+  end
+end